auth_service/requirements.txt
alembic==1.13.2 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.2.0 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==43.0.1 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.114.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.0 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.8 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.5 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.23.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.5.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic[email]==2.9.1 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
pytest-async==0.1.1 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.3 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
redis==5.0.8 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.34 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.38.5 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.30.6 ; python_version >= "3.12" and python_version < "4.0"
